+++ cluster_utils/__init__.py
@@ -11,7 +11,6 @@ from .client import (
     read_params_from_cmdline,
     save_metrics_params,
 )
-from .job_manager import grid_search, hp_optimization
 
 # The version is set based on git at install time, so we get it from the metadata of the
 # installed package here.  If this file is imported without the package being installed,
@@ -20,8 +19,6 @@ with contextlib.suppress(importlib.metadata.PackageNotFoundError):
     __version__ = importlib.metadata.version(__package__)
 
 __all__ = [
-    "grid_search",
-    "hp_optimization",
     "announce_early_results",
     "announce_fraction_finished",
     "cluster_main",
