+++ cluster_utils/server/__init__.py
@@ -1,0 +1,4 @@
+"""Server modules and scripts of cluster_utils.
+
+The `server` sub-package contains everything that is only needed for running the server scripts (e.g. `grid_search`/`hp_optimization`) but is not needed for the job scripts.
+"""
