+++ pyproject.toml
@@ -32,7 +32,14 @@ dependencies = [
 ]
 
 [tool.setuptools]
-packages = ["cluster_utils", "cluster_utils.base", "cluster_utils.client", "cluster"]
+# TODO: move packages to src/ and remove this entry
+packages = [
+    "cluster_utils",
+    "cluster_utils.base",
+    "cluster_utils.client",
+    "cluster_utils.server",
+    "cluster",
+]
 
 [tool.setuptools_scm]
 # Can be empty if no extra settings are needed, presence enables setuptools_scm.
